feat(jammer): Added support for adding songs to a playlist
Added the ability to add songs to a playlist using the 'playlist' command.
The user can now specify the name of the playlist and the songs they want to add by providing the names of the files or directories containing the songs.
The program will then create a new playlist if one with the specified name does not already exist, and add the specified songs to it.